+++ Tests/UnitTests/TestIssues/TestIssue41.cs
@@ -1,36 +1,32 @@
 ﻿// Copyright (c) 2020 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
-using System;
 using GenericServices.PublicButHidden;
 using GenericServices.Setup;
 using Tests.Dtos;
 using Tests.EfCode;
 using TestSupport.EfHelpers;
 using Xunit;
-using Xunit.Extensions.AssertExtensions;
 
 namespace Tests.UnitTests.TestIssues
 {
-    public class TestIssue43
+    public class TestIssue41
     {
         [Fact]
-        public void TestThrowExceptionIfILinkToEntityIsOwnedType()
+        public void TestThrowExceptionIfDtoPropertiesAreAllReadOnly()
         {
             //SETUP
             var options = SqliteInMemory.CreateOptions<TestDbContext>();
             using (var context = new TestDbContext(options))
             {
                 context.Database.EnsureCreated();
-                var utData = context.SetupSingleDtoAndEntities<DtoLinkedToOwnedType>();
-                var service = new CrudServices(context, utData.ConfigAndMapper);
+                var utData = context.SetupSingleDtoAndEntities<NormalEntityAllMarkedReadOnlyDto>();
 
                 //ATTEMPT
-                var dto = new DtoLinkedToOwnedType();
-                var ex = Assert.Throws<InvalidOperationException>(() => service.UpdateAndSave(dto));
+                var service = new CrudServices(context, utData.ConfigAndMapper); ///
 
                 //VERIFY
-                ex.Message.ShouldEqual("The class Address of style OwnedType cannot be used in Update.");
+                
             }
         }
     }